pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aimo-sampler"
version = "0.1.0"
description = "AI-powered audio sampler using natural language"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AIMO Team"}
]
keywords = ["audio", "ai", "music", "sampler", "musicgen", "generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.35.0",
    "audiocraft>=1.2.0",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "soundfile>=0.12.0",
    "pydub>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
ui = [
    "gradio>=4.0.0",
]

[project.scripts]
aimo = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
